*bugfix: lanv2 did not work for "long double" *example flens/test/geev.cc now uses "long double"

+++ flens/test/geev.cc
@@ -12,7 +12,7 @@ using namespace std;
 using namespace flens;
 
 // typedef mpf_class   T;
-typedef double   T;
+typedef long double   T;
 
 int
 main()
